Eliminate Var leaks in ClassField
Summary:
Fixes https://github.com/facebook/pyrefly/issues/533

It is always a potential data race to write a `Var` into a binding's
answer, because now all threads can see it and the first one that tries
to pin it wins; if there are any inconsistencies downstream that can be
triggered through resolving exports, we can get nondeterminism.

In most cases it turns out that these kinds of races are pretty rare
in practice, but ClassFields can have particularly large fanouts, and the
data race is more likely. We need to prevent it - this diff does so by
forcing all vars.

I've left a TODO to think about whether we could improve inference using
a scheme similar to first-use inference for assignments. That seems
possible, but I suspect it's harder to do well using the current approach
so I intend to defer it for now; I think it's possible that cycle breaking
will move our architecture in a direction that makes this easier to do
cleanly.

----

I'm a little fuzzy on why this affected a completion test, but I have pretty
good confidence that the change is reasonable, it's a one liner with
pretty well-understood behavior.

cc kinto0, I'm inclined to ship this without diving into the LSP but this
does seem pretty strange to me

Reviewed By: ndmitchell

Differential Revision: D77031603

fbshipit-source-id: 56c8de34fc16923475a0ec73bd72b4a9d62a72a4