Add a test showing class field var leaks are problematic
Summary:
This test shows that class fields with Vars behave pretty oddly, since the first use
we see pins. That's obviously potentially surprising, although first-use pinning is
okay when done well.

But we don't have actual first-use pinning for ClassField, so this test is essentially
showing it can lead to nondeterminism:
- if the current file contains a use and type checks first, we'll pin it to the use;
  if it gets exported first we'll pin it to Any instead
- if the current file has multiple uses and they can be triggered in different orders
  based on exports, we might pin to different things depending on thread races.

Reviewed By: ndmitchell

Differential Revision: D77031606

fbshipit-source-id: a1624b95addc27bc6830409dedb3bdbd190fcb5a