Refactor find_one_part_prefix
Summary: This code had unsafe indexing, loops, allocation and more. Change to a few if conditions, which I think precisely preserves the behaviour? I'm not 100% sure though, and not 100% sure the behaviour was correct before (the second prefix check looks redundant?). Please review carefully, and ideally be the person who wrote this originally.

Reviewed By: connernilsen

Differential Revision: D76969153

fbshipit-source-id: 52827412552b8d6f4a8e1b945a677155922451d9