Add a new cycle test based on Issue 528
Summary:
We have an opten issue tracking a loop example where we produce a recursive
Var for a cycle, pin it at some point in the cycle, then get back to the
original binding where we broke the cycle and when we solve the original
binding we discover that the type is inconsistent to what got pinned.

This kind of thing is several kinds of not great:
- The user-facing behavior is quite nasty any time we hit this error
- It is related to determinism issues; this particular example is probably
  deterministic, but we've observed the same kind of error popping up
  nondeterministically in some projects.
- It indicates a bug in our actual logic, since this code is actually okay.

Importing it to track the issue in a way where I can:
- easily observe any behavior changes if we change how cycles work (which
  I've been prototyping; this might affect the behavior although I
  don't believe it will fix the root cause)
- easily see any behavior changes if I alter how scope works, which I
  suspect is the proper fix here

----

I know what's happening here but am not yet sure what to do about it; it's
related to multiple mid-size projects I am considering trying to tackle
more or less right now, and it's not clear that a quick fix would really be
helpful if the quick fix is too narrowly tied to this exact snippet of code.

Basically, the root cause is a combination of related things, several of
which are bugs and all of which have to occur for this example to break:
- the `for` loop header is considered outside the loop (which is reasonable)
- then that means our auto-speculative-phi logic automatically makes `item` a
  speculative phi inside the loop; this is unreasonable for two reasons:
  - first, there's no assignment, and I think speculative phi is probably
    something we should just avoid when there's no assignment
  - second, even if there were assignment, the `for` header actually reruns
    at the top of every loop, so any type at the end of the loop isn't
    really relevant
- then, the `isinstance` check winds up producing a subtype check, and without
  better solver isolation (today, the solver bangs away at a global mutex, and
  there's no plumbing yet to do anything else) this will pin the recursive var...
  this is also probably a bug, `isinstance` checks are *narrowing* operations
  that are speculative so it never feels morally right to pin a type.
  - this needs to be fixed, it's essentially the same as one of the major issues
    we have with overload resolution, but it will likely be a nontrivial
    refactor
- when we get to the end, the `isinstance` narrowing is no longer live (that's
  entirely correct!), so the phi finally solves itself to a type that doesn't
  depend on `T` (I'm a little fuzzy where the `Iterable[T]` is coming from,
  but I currently think that detail is less relevant to bigger architectural
  concerns; that part doesn't mismatch anyway. I may be wrong about this detail
  being less important.)

Reviewed By: ndmitchell

Differential Revision: D77037445

fbshipit-source-id: e9887a3fb94fda47ed1e93ffe8b80ffb676869eb