Add del testcase for TypedDict
Summary:
There is a TODO that we need to synthesize `__delitem__`, but I have tested all the cases I can think of for this and they all work.

It seems that this was already handled here? D71197974
I just added more tests to verify if there are any missing corner cases, but it seems everything is covered?

Reviewed By: ndmitchell

Differential Revision: D77038037

fbshipit-source-id: 79438f9d681227653deb0b0f5f836952e46cd39a