Fix non-determinism in autocomplete integration test (#534)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/534

The non-determinism is likely coming from that there are multiple matches, and one match is on the edge of being either matched or not matched.

This diff addresses the issue by running the test on an extremely long name instead, so that the fuzzy matcher will always match it with high confidence.

Reviewed By: ndmitchell

Differential Revision: D77039902

fbshipit-source-id: af9647fcd492d248f76afb8bd1101ebe1a9dd306