Move `class_tparams` to the top of `class_metadata.rs`
Summary:
This concept is much simpler and more primitive than `ClassMetadata`,
so it makes more sense to have it defined above rather than buried
in the middle of the rest of the metadata logic.

It likely ought to be moved elsewhere eventually, but I'm not ready
to do that yet; it's a little tricky since both tparams and metadata
share a reliance on `BaseClass`, which makes moving it out bring up
a lot of bikeshedding questions I'm not ready to deal with right
now.

Reviewed By: ndmitchell

Differential Revision: D77126630

fbshipit-source-id: cada0cb22807616c1230a1f9add2ca4af1cd59d6