Move `Class` implementation
Summary:
I find it a little hard to skim code where all the important type defs
(`Class` itself is not important at all, it's just an arc around the
actual data) live below the logic defined in terms of them.

Just reorganizing things so it's easier for me to skim, as I need to
make some refactors as part of an effort to eliminate cycles in
classes.

Reviewed By: ndmitchell

Differential Revision: D77126631

fbshipit-source-id: 156a4b21f6f9f904ce92c5ca2300aa0516be5ddc