Move `TArgs` and `Substitution`
Summary:
These don't belong in the `class.rs` module; they are used in `Forall` types,
including generic type aliases, as well. So the concepts are more general than
the class code.

Having them in the class module is contributing to difficulty decoupling
them to eliminate cycles in highly complex class hierarchies.

In addition, I think it has always been pretty weird for `TParams` and
`TArgs` to be defined far away from one another, they are very closely
related concepts and it is pretty helpful to have them come one
after another, along with `Substitution` which is essentially the operation
that relates them.

Reviewed By: ndmitchell

Differential Revision: D77126627

fbshipit-source-id: 76270ec6acdd4219b42260a68e172cd288283d9f