Delete `Class::as_class_type`
Summary:
This function cannot exist in a world where Pyrefly handles
legacy type parameters in a manner that inspects types, and
also supports class hierarchies cyclic enough for the current
class to appear in the type parameters of its own base classes
(which is mandatory - codebases we must support do this).

It needs to instead live on the solver, so that it can redirect
through bindings in order to be lazy enough not to produce
cycles that break the model of the class hierarchy.

This diff moves it.

Reviewed By: ndmitchell

Differential Revision: D77126632

fbshipit-source-id: fe5630a4be553a0aa309705011e9d61467ad2ce7