Store TParams pointer inside `TArgs`
Summary:
This diff makes the `TArgs` aware of the `TParams` it pairs
against.

This provides an extra layer of protection against creating an
invalid TArgs which is nice, but more importantly it makes `TArgs`
know enough to display itself, which is what I'll work on next;
this was blocking me from eliminating the coupling between `Class`
and `TArgs.

Reviewed By: ndmitchell

Differential Revision: D77126626

fbshipit-source-id: 018bda269c96294976a5bb05c830ba47ba0eeb5d