Make `TParams` an `Arc` in `ClassInner`
Summary:
In order to eliminate some cycles that are causing major
problems in sufficiently complex class hierarchies (in particular,
hierarchies where a class is generic in itself), I want to
extract `TParams` from `Class`.

That's a substantial refactor if done all at once, so I'll split
it into easy stages. The end result is going to require `TParams`
to be a dedictated binding that produces a result, which means we're
refactoring toward a world where there's a method on `AnswersSolver`
to get an `Arc<TParams>`.

This diff begins the type tetris to make that happen: first, we need
to put it in an arc on `ClassInner`. Then we'll gradually rewrite the
apis to look the way we need, and we'll finish by creating the new
binding kind at which point the cycle should be eliminated.

Reviewed By: ndmitchell

Differential Revision: D77126634

fbshipit-source-id: ce067ab995bfb3b5121fa7df7d08711eb5d8433d