Arc up all the TParams
Summary:
To have a reasonable state of the actual semantics at the end of
my refactoring to decouple TParams and Class, I really want the TArgs
to be able to store a copy of the params they coupled against.

Why? It turns out TArgs are pretty much useless without the params because they
aren't printable (which means you can't use them in an error message, which
means you can't type check anything at all with them, unless the params are
present).

To me, this strongly suggests that TArgs should know the parameters, that
way it is display-able and therefore generally usable. To do that without
unnecessary clones, we want to Arc up the `TParams` everywhere (in many
cases they'll be coming from a new binding in the end state of this refactor).

Note that we have to implement a nullary `VisitMut<Type>` here... this is
exacty the same nullary `VisitMut` we already implemented on `Class`, needed
for exactly the same reason; it's not surprising I might need it, since I'm
specifically aiming to get the `TParams` *out* of `Class`!

Reviewed By: ndmitchell

Differential Revision: D77126624

fbshipit-source-id: 97472ea056a7d567e3f0b6da570c5aaee2b5f39d