return CommandExistStatus::UserError instead of Error when encountering issues when loading from pyproject
Summary: This currently returns an error if we can't find a mypy nor pyright config in pyproject.toml. In theory, this doesn't happen since the only entrypoint to this function guarantees that we already know that the mypy or pyright config exists, but it's possible that `init.rs` will change in the future and this will prevent panics.

Reviewed By: MaggieMoss

Differential Revision: D76851746

fbshipit-source-id: 47369ec59eb08d27b3bd12857fde6619513f5c62