Eliminate `Class` as an input to `Substitution`
Summary:
We don't need it anymore - the `TArgs` already know what params they map
to. This was a bit smelly anyway - we actually perform substitutions on
Forall types occasionally, but we were unable to use the `Substitution` helper
and instead use raw maps of quantifieds which was needed because `Substitution`
was inappropriately coupled to `Class`. I'll clean that up next.

Reviewed By: rchen152

Differential Revision: D77161655

fbshipit-source-id: df01e778b90619207bb8dbcbed416b6cb65f718d