Use `Substitution` in `Forall`, also rename
Summary:
No reason to duplicate low-level fiddly logic in `Forall`.

Also, this function is better named `apply_targs`, which is less
ambiguous about what's going on than `subst` is (`subst` is okay,
but tells me less - what are we substituting for what? Have to read
the code to find out).

Reviewed By: yangdanny97

Differential Revision: D77161652

fbshipit-source-id: a0a552a67bf0b662e0190c82469651c8b3d58229