Synthesize a __del_item__ method for typeddict
Summary:
`del` vs `__delitem__` do not yield the same behavior.

while `del` does implement the correct logic based on the typeddict's fields, since we don't have a synthesized method, using the dunder method directly will fallback to typeshed and reject everything.

I noticed that pyright doesn't complain about any deletion. The signature for the function invoked in pyright is of the form `(function) def __delitem__(k: str) -> None`

I think we could be more restrictive by not allowing deleing required keys so that it can look similar to `del`?

The logic is is just constructing overloads based on the fields we encounter, but we skip required fields.

I think the error messages can be improved, but this is probably a separate problem.

Reviewed By: stroxler

Differential Revision: D77081695

fbshipit-source-id: a80d24591be0eb8617459ac137897c49bcbf5fcc