Add Partial type variant
Summary:
We should add a new `Partial` variant so we can use it to synthesize the update method for TypedDict.

Pyright does something similar.

Consider this class

  class C(TypedDict):
    x: int
    y: int

Pyright will synthesize:

  def update(
    __m: Partial[C],
    /
  ) -> None: ...

This annotation will allow us to specify `Partial[C]`, without having to specify that `total=false` for `C`.

`Partial[C]` represents a type with any combination of keys from `C`. Thus, it's compatible with  `{"x": 1}`.

My goal is to add a variant to represent those semantics, and then to use it as an annotation in the update method.

Reviewed By: rchen152

Differential Revision: D77181950

fbshipit-source-id: 64e8fa5110224a76c831e2db2c1146847f92fb37