Prepare `stdlib` for a world where `TParams` lives outside `Class`
Summary:
Beginning this diff, I'm going to start prepping all the call-sites of
`Class.tparams()` to stop relying on it; instead, just a few central places
will rely on `Class.arc_tparams`. This will make the key change that actually
migrates `TParams` to its own binding much easier to roll out, since only the
few central places will need to change instead of dozens of call sites that
might not have access to an `Answers` or `AnswersSolver`.

This diff makes the change for `stdlib`: we make the `state.rs` code (which has
access to `lookup_answer`) responsible for calling `class.arc_tparam`; it will
eventually need to instead look up the binding when necessary. The stdlib module
itself just understands that these are two bits of data and both are needed.

Reviewed By: ndmitchell

Differential Revision: D77161650

fbshipit-source-id: 1948657525c3b4e678909b9d5f9389c584b1be31