Wire an optimization into existing logic
Summary:
In order to reduce the need for extra bindings, we'd like to conservatively
try to skip moving `TParams` into a separate binding that deals with legacy
tparams for the (very common) case where there are no generic bases.

To verify that this works independently of the refactors involved in splitting
bindings, I'm first implementing this shortcut in-place in the existing
`BindingClass`: we can skim the bases in the bindings phase, and just pass
empty `bases` whenever there's no chance of legacy type parameters being bound.

Reviewed By: ndmitchell

Differential Revision: D77181969

fbshipit-source-id: 0eee518741557afcc73b13fed418f5bba3f132e7