Renames
Summary:
I realized that the `compute_tparams` function, which is now visible
as a public helper on `AnswersSolver`, has a name that is too ambiguous;
you can't see from the method calls that it lives in `alt/class`, so
it really needs to have a name that makes it clear this is about *class*
tparams, not *general* tparams.

In the same vein, I though it would be better to use `validated_tparams` for
the general helper (which applies to function and generic alias tparams in
addition to just class tparams) so that the name is consistent with other
names, and makes it clear that there is some validation involved here.

Reviewed By: ndmitchell

Differential Revision: D77181970

fbshipit-source-id: 1b3477ccab3d491b29636ade9c4bbfbc53f679bb