Split `TParams` into its own binding
Summary:
This commit finishes the breaking of cycles that cause us problems
in class hierarchies that include self-referential generics: we
factor the legacy-type-param-handling part of `TParams` computation into
a separate binding.

To minimize wasted lookups, we will precompute the tparams based just
on scoped type variables whenever we are confident (by syntactically
analyzing the bases) that there are no legacy type variables to worry
about.

Reviewed By: ndmitchell

Differential Revision: D77227322

fbshipit-source-id: d640fb25daf12af69eafb1f75b1f02106f9dcede