Make renamed import be the boundary for global find refs
Summary:
For find-references with renamed imports, we want the following behavior. Suppose we have

```
# a.py
foo = 3

# b.py
from .a import foo as bar
bar
```

Find-ref of `foo` from `a` should include the `foo` from `a` and the import from `b`, but not those two `bar`s. Similarly, find-ref of `bar` in `b` should only find the two in `bar`, and not include any of `foo`.

Why? While it can be debatable for find-refs, this is the only behavior that makes sense for rename, which should have the same behavior as find-ref.

This diff implements the rule using the new ranges we start tracking earlier in the stack. We introduce a special mode of find-ref where we don't jump through the renamed import, so that these renamed imports act as a firm boundary. In this way, local find refs won't jump through the import, and the results will be limited to the local file. In addition, in the indexing step, we will separately index these renamed exports, so that that are correctly accounted.

Reviewed By: kinto0

Differential Revision: D77196605

fbshipit-source-id: 80b74ca8dfac02aa5d83c95287e2e7a4ee8c1f27