Add a test case of cyclic nondeterminism in class metadata
Summary:
We have a bug in our class metadata logic that is doubly affecting us
when there are complex class hierarchies with cycles through generics:
- Sometimes, we are detecting a cyclic MRO that isn't actually cyclic
- Moreover, this behavior seems to be nondeterministic - in the tests
  added here, if I try to resolve `A().x` first it is okay but if I
  try to resolve `B().x` first we fail.

I'm almost sure this is another example of cycle-breaking nondeterminism;
when we try to resolve `A().x` first, the entrypoint to the class metadata
will be `A` and that's where we'll break.

I'm fuzzy on what would actually cause this. The MRO isn't actually
cyclic; maybe there's a bug in the MRO itself, but we've also steadily
pushed more stuff into the class metadata so that it is far more than an
MRO today. It might be that some other part of the metadata is cuasing
the problem. These test cases will hopefully help us root cause the issue.

Reviewed By: ndmitchell

Differential Revision: D77232060

fbshipit-source-id: 6d40c0c8d49ac9136ab592e0a6d7038c61ffa229