Explicitly name the two cases when solving `ReturnType` binding
Summary:
organize_all_the_code

The solving logic for `ReturnType` binding is in need of some cleanup: Conceptually, there are actually multiple different cases we want to consider when solving return type for a given function, but the data structure `Binding::ReturnType` does not make that clear at all: it's just one single bag of data fields put together, some of which are relevant for some cases, and some of which are relevant for some other cases. This is a classical example of redundancies caused by using product type to model something that's better modeled with sum type.

To untangle the situation a little bit, this diff proposes we have `ReturnType` contain a little enum explicitly differentiating all the cases. At the moment, I think two cases are the most relevant:
- One case  (`ShouldValidateAnnotation`) deals with the situation when a return annotation is explicitly provided. In this case, we currently always want to validate that annotation against implicit returns, and use that annotation for downstream callers.
- Another case (`ShouldInferType`) deals with the situation when a return annotation is not explicitly provided. In this case, we currently always want to infer return types based on both explicit return/yield and implicit return, and use that inferred type for downstream callers.

Later down the stack we'll introduce more cases to this list (see D77250008) . This diff makes that introduction easier by refactoring the pre-existing logic.

Reviewed By: stroxler

Differential Revision: D77250004

fbshipit-source-id: 5b8728e0e6e62d1f1d1a9a1fc57465b0be920774