Only keep full `yields` and `yield_froms` info when inferring return types
Summary:
justmovingthingsaround

We only need the full type information for all `yield`s/`yield from`s in the function when inferring return types. If we are just validating the return annotation, all we need is 1 bit of info telling us if the function syntactically contain any `yield`/`yield from`, in which case we could just preserve a boolean flag.

This change helps us take fields out of `ReturnType` binding which enables D77250005

Reviewed By: stroxler

Differential Revision: D77250007

fbshipit-source-id: cb68b6817c51819896d6e7886124d0ceefcf7364