Introduce two new kinds of return type variant: `ShouldTrustAnnotation` and `ShouldReturnAny`
Summary:
This diff introduces two more variants to `ReturnTypeKind`:
- `ShouldTrustAnnotation` means the function is annotated and we should just blindly trust the annotation without validating it against anything. This behavior is useful when function is explicitly annotated and `untyped-def-behavior` is not set to anything other than `check-and-infer-return-type`.
- `ShouldReturnAny` means the function is not annotated and we should just assume the function returns `Any` (aka Pyre1 behavior). This behavior is useful when function is not annotated and `untyped-def-behavior` is set to `skip-and-infer-return-any` or `check-and-infer-return-any`.

These two cases are currently "emulated" by two other kinds of bindings (i.e. `Binding::AnnotatedType` and `Binding::Type(Type::Any)`), but that's incorrect since those bindings lack necessary return-annotation-to-return-type transformation (see D77250003). In this diff, we introduce the two additional `ReturnTypeKind` to bake in the return-annotation-to-return-type transformation behavior. In D77250003, we'll connect the behavior to prod.

Reviewed By: stroxler

Differential Revision: D77250008

fbshipit-source-id: ec55b1cd2926890ea6891ff8c6a1e6f024410019