Only keep full `returns` info when inferring return types
Summary:
imjustmovingshitaround

We only need the full type information for all `return`s in the function when inferring return types. If we are just validating the return annotation, all we need is 1 bit of info telling us if the function syntactically contain any `return`, in which case we could just preserve a boolean flag.

This change helps us take fields out of `ReturnType` binding which enables D77250005.

Reviewed By: stroxler

Differential Revision: D77250006

fbshipit-source-id: ed688c800e6b71f59ae6b49cc1daf1644ea99ed3