Clean up MRO binding: just use metadata
Summary:
The ClassMetadata binding can safely be made *upstream* of the MRO;
in fact it already was because an edge case in our typed dict handling
required injecting an extra base class based on the `is_typed_dict`
field.

This is okay because the behavior we needed to avoid was that the MRO
calculation itself just can't be in the same cycle as class-metadata and
expression-to-type inference; it's fine for it to depend on the metadata.

This simplifies the logic, and avoids duplicating the hook for the typed
dict edge case (which worried me because that kind of complexity where
the bases change based on logic buried in a multi-hundred-line function
is not practical to duplicate without bugs, there has to be just one
source of truth).

Reviewed By: ndmitchell

Differential Revision: D77301477

fbshipit-source-id: b28c5c00bf5e23f6877666a682ff76c67ed64e44