Simplify MRO computation
Summary:
The MRO computation has no need to be dealing with `ClassMetadata` as a whole,
it is all about MRO and nothing else; as we grew `ClassMetadata` to include many
other fields, we kept pushing it all down to eliminate a new vec allocation, but
this is happening only once per class so it's probably not worth the lack of clarity
in code structure.

More importantly, though, I now believe that we need to extract the MRO into its
own binding; the metadata is used sometimes in `infer_expr`, which means that
it is potentially cyclic. That is actually okay, particularly if we eliminate
nondeterministic cycle breaking, because the cycles don't actually cause problems
in most of the metadata.

But they cause major problems once we encounter them in the MRO - if we encounter
a cycle actually induced by other metadata in the MRO, we will conclude that the
MRO is cyclic, and we'll both declare the class invalid and ignore the entire
ancestor chain. This is not okay, hence my desire to extract MRO into a separate
binding (expr_infer does *not* depend on the MRO except in truly cyclic cases -
like trying to use an attribute defined on a class in the cycle as a base class -
that would actually crash the runtime).

This diff is both a simplification of existing logic and a step toward separating
the metadata from the mro.

Reviewed By: ndmitchell

Differential Revision: D77281717

fbshipit-source-id: 37579eaa757d1a0bf3eb558af0e7c86d8ff05c3b