Implement textDocument/prepareRename support
Summary:
The implementation is very straightforward, and basically a wrapper around `Transaction::identifier_at`.

This is optional, so why do we have to implement this? Well, this is necessary for VSCode and potentially other LSP clients to choose which rename provider to invoke when there are multiple rename providers. Although pylance will disable itself when pyrefly is active, we can run into conflicts if there are other DSL language server active that also acts on python files.

Reviewed By: kinto0

Differential Revision: D77245353

fbshipit-source-id: e75c916bb47e50c767b2e0f09f64a083a790dd9b