prompt user to suppress errors if there are 100 or less errors
Summary:
Title, prompt user to suppress their errors if there's 100 or less errors. This help set the project up for a clean type check.
I picked 100 as an arbitrary number that I thought was small enough to allow errors to be suppressed.

For cases where there's more errors ,I plan to prompt the user to turn on typechecking for specific directories instead.

Reviewed By: MaggieMoss

Differential Revision: D77156268

fbshipit-source-id: b7f5f8d1ee0ca9cdddb2f7a6d485c59ae7e0b5ac