Wire in conda environment
Summary:
This diff and the previous add functionality that enables users to specify a conda environment name that will be used, even when not activated, unless:
- a venv is active
- a conda env is active
- a venv is present in the project directory

This diff finishes the functionality for using conda environments, even if they're not active. In the `configure` function, we add the ability to query conda for an environment. We also update the LSP way of overriding a config to match what we do when a `python-interpreter` is set, though we may be removing that in the future.

Reviewed By: yangdanny97

Differential Revision: D77247852

fbshipit-source-id: 1fefe95d0bd06a0424c2bc7d53c5b2890e4a609d