Factor `AnswersSolver` out of `answers.rs`
Summary:
Having all the Answers and Solutions logic - which mostly
has very little to do with solving answers, with a few
exceptions - living in the same module as the core of the
answers solver was making it harder for me to add the extra
state needed for cycle breaking.

I think this separation of concerns is probably good hygiene
anyway, but my main motivation is that it will make adding more
logic to `Cycles` and `get_idx` a lot more manageable.

Reviewed By: ndmitchell

Differential Revision: D76948129

fbshipit-source-id: aa999afa91323122cb40c98346fc097810d7b47d