Create a more flexible (private, for now) interface for `Calculation`.
Summary:
I believe that deterministic cycle breaking will require the ability to use some strategy
other than immediately recording a recursive placeholder result when a cylce is detected.

With this in mind, I want to work on decoupling the detection of a cycle versus the
recording of a recursive placeholder to break it.

This diff does so while keeping the same public interface; the `calculate` and
`calculate_with_recursive` methods are now built out of simpler APIs that I believe
can be extended to support deterministic cycle breaking.

Later in the stack I expect to need a (slightly more powerful variation of) the new
interface in `get_idx`.

----------

Edit: just pushing down notes from the top of the stack: implementing this on Pyrefly
today appears to roughly halve the nondeterminism on an example project. It's known
that we still leak recursive `Var`s in ways prone to data races; my current theory is that
most of the remaining nondeterminism comes from that; any reasonable solution I've been
able to think of so far would build on top of cycle-breaking logic.

Reviewed By: ndmitchell

Differential Revision: D76948138

fbshipit-source-id: 6e74ff378f32aaa22fc0d972b1afcffb55e9691e