Fix up test_find_module_prefixes (#576)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/576

This code finds module prefixes in the order they appear on disk. Often files written in one order will read in that order, but there is no guarantee, and it's file system dependent. To make sure these tests are deterministic on all file systems, sort the result.

Note that before the result is used in LSP we sort it anyway, so it doesn't need sorting in the real method.

Reviewed By: grievejia

Differential Revision: D77373942

fbshipit-source-id: fc333f02c1c4f48aaef81c3032ae5be051f8936b