Propagate TypedDict's update type hint to expression type inference
Summary:
We are adding a new Partial type that wraps TypedDicts to be able to type "udpate". For context, see D77181950.

After constructing the partial type when synthesizing an "update" method, we want to propagate the type hint to the expr infer stage.

However, to do so, we must add some special handling here https://www.internalfb.com/code/fbsource/[b4a4e4f927f1]/fbcode/pyrefly/pyrefly/lib/alt/call.rs?lines=663

to do this, we want to know that the method we are looking at at this point in the code, is an update method belonging to a TypedDict.

This diff allows the type hint to propagate.

Reviewed By: rchen152

Differential Revision: D77336201

fbshipit-source-id: c641a0c2b09bd3cf05e5a65a810c0ffe1bc2f1b7