Add extra cache clearing functionality for LSP
Summary:
# Problem
While Xavi was working on his intern project, he was running into issues where a workspace configuration change wasn't being reflected in the Python configuration being used.

After looking into it, we realized configurations weren't being reloaded on a workspace change. This shouldn't be happening, and in any project with an on-disk configuration, we couldn't replicate the issue. `State::invalidate_config()` should be invalidating the caches, and we saw that the file was being iterated over, so we deduced the config wasn't being reloaded.

It turns out that there are two caches in `standard_config_finder` that aren't cleared on `State::invalidate_config`. We were able to figure out that for any projects without a config file, the caches there wouldn't regenerated the config.

## Rationale

This makes sense for CLI checks, since once we're able to generate a synthetic config, it doesn't ever change unless a config file is added on disk, which we pick up properly.

Pyrefly's LSP has extra hooks in the `configure()` functions that update based on state changes from VSCode, so this approach doesn't work for projects without configs.

# Solution

The solution I've come up with so far is to pass `standard_config_finder`'s caches through to the `ConfigFinder`, and have `ConfigFinder::clear()` handle clearing them for us.

## Alternative Approaches

I really wanted to instead turn the finding functionality of the `ConfigFinder` into a trait, but it turns out we can't do that with the generic functionality `ConfigFinder` is designed around right now. There might be a way to do this that I'm just too Rust-novice to know about right now, so open to alternatives. For now, we do this with an extra `clear_extra_caches` function that's passed a long with references to the caches the `standard_config_finder` creates.

The other approaches I've considered as well are:
1. We could get rid of the generic functionality there, since it's unlikely we'll be loading other config types. Then we could move `before`, `fallback`, and `clear` into the trait and have `StandardConfigFinder` be an implementation of that trait.
2. Instead of passing a clear caches function to `ConfigFinder`, we could have the caches be static mutexes, which we only clear in the IDE, but then the IDE's implementation is tied to `standard_config_finder`, which is probably not a great approach. The good part about this though is that CLI checks won't ever clear those caches, so we won't redo the work there where it won't ever be needed.

Any thoughts on the approach? Any alternatives we should try before committing to this?

Reviewed By: ndmitchell

Differential Revision: D77417749

fbshipit-source-id: 8e7a37449a105b0bb2aa7f5e70dea16760a938f6